refactor: rename Eval -> PreDiff, pre_eval -> pre_diff, eval -> diff_if
