+++ crates/sauron-core/src/dom.rs
@@ -16,7 +16,7 @@ mod task;
 use cfg_if::cfg_if;
 
 cfg_if! {if #[cfg(feature = "with-dom")] {
-    pub use application::{Application, Measurements, Eval, eval};
+    pub use application::{Application, Measurements, PreDiff, diff_if};
     #[cfg(feature = "custom_element")]
     pub use web_component::{register_web_component, WebComponent, WebComponentWrapper};
     pub use dom_patch::{DomPatch, PatchVariant};
