+++ crates/sauron-core/src/dom/application.rs
@@ -2,83 +2,11 @@ use crate::dom::Cmd;
 use crate::vdom::Node;
 use mt_dom::TreePath;
 use std::fmt;
+pub use pre_diff::{PreDiff,diff_if};
 
 
-///
-pub struct Eval{
-    expr: Box<dyn Fn() -> bool>,
-    children: Vec<Eval>,
-}
+mod pre_diff;
 
-impl fmt::Debug for Eval{
-
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result{
-        write!(f, "({}", (self.expr)())?;
-        write!(f, ",[")?;
-        for child in &self.children{
-            child.fmt(f)?;
-        }
-        write!(f, "])")?;
-        Ok(())
-    }
-}
-
-
-impl Eval{
-
-    /// new
-    pub fn new(val: bool, children: impl IntoIterator<Item = Self> ) -> Self {
-        Self{
-            expr: Box::new(move||val),
-            children: children.into_iter().collect()
-        }
-    }
-
-
-    ///
-    pub fn none() -> Self {
-        Eval{
-            expr: Box::new(||false),
-            children: vec![],
-        }
-    }
-
-    ///
-    pub fn traverse(evals: &[Eval]) -> Vec<TreePath> {
-        let root = TreePath::root();
-        if evals.len() == 1 {
-            Self::traverse_recursive(&evals[0], root)
-        }else{
-            Self::traverse_list(evals, root)
-        }
-    }
-
-    ///
-    fn traverse_list(evals: &[Eval], current: TreePath) -> Vec<TreePath> {
-        let mut paths = vec![];
-        for (i,eval) in evals.iter().enumerate(){
-            let more_paths = eval.traverse_recursive(current.traverse(i));
-            paths.extend(more_paths);
-        }
-        paths
-    }
-
-
-    fn traverse_recursive(&self, current: TreePath) -> Vec<TreePath>{
-        let mut paths = vec![];
-        if (self.expr)(){
-            paths.push(current.clone());
-        }
-        let more_paths = Self::traverse_list(&self.children, current);
-        paths.extend(more_paths);
-        paths
-    }
-}
-
-/// evaluate check
-pub fn eval(val: bool, children: impl IntoIterator<Item = Eval>) -> Eval{
-    Eval::new(val, children)
-}
 
 /// An Application is the root component of your program.
 /// Everything that happens in your application is done here.
@@ -104,9 +32,11 @@ where
     where
         Self: Sized + Clone + 'static;
 
+    #[cfg(feature = "pre-diff")]
+    /// an optimization solution.
     /// pre evaluate the expression to determine
     /// whether to diff the nodes
-    fn pre_eval(&self, _other: &Self) -> Option<Vec<Eval>> {
+    fn pre_diff(&self, _other: &Self) -> Option<Vec<PreDiff>> {
         None
     }
