+++ examples/counter/start.sh
@@ -1,3 +1,3 @@
-wasm-pack build --release --target=web &&\
+wasm-pack build --release --target=web -- --features "sauron/pre-diff" &&\
 
 basic-http-server -a 0.0.0.0:4000
